feat: 일정 생성 시 pace, distance 정수형으로 수정

+++ src/main/java/com/youthexpedition/azit/modules/crew/application/port/in/command/CreateScheduleCommand.java
@@ -16,8 +16,8 @@ public record CreateScheduleCommand(
         String meetingSpot,
         Double latitude,
         Double longitude,
-        Double distance,
-        Double pace,
+        Integer distance,
+        Integer pace,
         Integer maxParticipants,
         String description,
         List<String> supplies
@@ -25,7 +25,7 @@ public record CreateScheduleCommand(
     public static CreateScheduleCommand of(
             Long crewId, Long creatorId, String title, RunType runType, LocalDateTime meetingAt,
             String placeName, String address, String meetingSpot, Double latitude, Double longitude,
-            Double distance, Double pace, Integer maxParticipants, String description, List<String> supplies
+            Integer distance, Integer pace, Integer maxParticipants, String description, List<String> supplies
     ) {
         return new CreateScheduleCommand(
                 crewId, creatorId, title, runType, meetingAt,
