+++ src/main/java/com/youthexpedition/azit/modules/crew/application/port/in/dto/CrewScheduleDetailResponse.java
@@ -24,9 +24,9 @@ public record CrewScheduleDetailResponse(
         @Schema(description = "일정 설명")
         String description,
         @Schema(description = "목표 거리 (km)")
-        Double distance,
+        Integer distance,
         @Schema(description = "목표 페이스")
-        Double pace,
+        Integer pace,
         @Schema(description = "최대 인원")
         Integer maxParticipants,
         @Schema(description = "현재 참여 인원")
